[fix] Refactor ProfileCard component
